irqchip/apple-aic: Only access IPI sysregs when use_fast_ipi is true
Starting with the A11 (T8015) SoC, Apple began using ARM64 sysregs for
fast IPIs and UNCORE PMC control within the AIC. On their earlier armv8
designs (A7-A10), accessing these sysregs results in a more or less
instant crash, which is quite a good indication they're not serving the
same purpose.

Restrict sysreg access within the AIC driver to configurations where
use_fast_ipi is true (targets using the apple,t8103-aic and apple,aic2
compatibles) to allow the AIC to function properly on these older
designs.

While at it, remove the IPI-always-ack path in aic_handle_fiq - if
we're able to reach and execute that path, we are running a fast
IPI-capable system and should be using one of the other compatibles,
anyway.

Signed-off-by: Konrad Dybcio <konrad.dybcio@somainline.org>