[wip, need to separate this] arm64: dts: apple: Add A9/X devices
Add DTS files for the A9/X SoC and the following devices based on it:
- iPhone 6S
- iPhone 6S Plus
- iPhone SE (2016)
- iPad 5
- iPad Pro (1st gen) (A9X)

The A9 SoC comes in two flavours: S8000 (Samsung) and S8003 (TSMC). While
there aren't any obvious software differences now, keep the DTs separate
to make it less of a pain to introduce chip-specific code paths later on,
should there be any. Even if there won't be, it's good to describe the
hardware properly, hence separate compatibles and mentions of the SoC in
the model property.

This does not however concern the A9X (S8001), which is only made in a
single fab.